add support for zip files + small improvements

+++ l10n_be_coda_advanced/models/account_coda.py
@@ -3,6 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo import api, fields, models, _
+from odoo.exceptions import UserError
 
 
 class AccountCoda(models.Model):
@@ -45,6 +46,9 @@ class AccountCoda(models.Model):
     @api.multi
     def unlink(self):
         for coda in self:
+            if coda.state != 'draft':
+                raise UserError(
+                    _("Only CODA File objects in state 'draft' can be deleted !"))
             coda.bank_statement_ids.unlink()
         return super(AccountCoda, self).unlink()
